use patch decorator for patching datetime.datetime
The problem with not using a patch decorator is that the datetime.datetime is
patched for the test AND for all other tests. I.e. it is a side-effect that is
valid globally for the rest of the execution of the program. I.e. all other
calls to datetime.datetime.now will be 01.01.1970. For example even pybuilder
will finally output that this is the date it was run at.

This can lead to a very nasty phenomenon where tests that NEED the patch but
DON'T have it initially due to a programming error will pass because the global
patch is still valid from a previous test. Depending on the order in which the
tests are executed they may randomly pass or fail. The order may be different
on different platforms and so the tests might misleadingly indicate that there
is an issue with the platform.

This commit uses the patch decorator instead which ensures that the original
datatime.datetime is put back into place after the test(s).