fix for dst switch bug
Alppaca is known to have a bug when switching to daylight savings time (dst).
I.e. at the beginning of winter when the clocks are set back an hour. It isn't
clear as to where this problem originates (alppaca itself or the underlying
apscheduler) and it is fairly difficult to reproduce reliably.

This patch changes the default time-zone of the underlying scheduler component
to UTC and schedules all credentials refreshes in utc too. It is to be taken as
a "wild guess" in the right direction.